+++ CHANGELOG.md
@@ -24,9 +24,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `Internal\Framework\Module\Configuration\Dao\ShopEnvironmentConfigurationDaoInterface`
   - `Internal\Framework\Module\Configuration\DataMapper\ProjectConfigurationDataMapperInterface`
   - `Internal\Framework\Module\Configuration\DataMapper\ShopConfigurationDataMapperInterface`
+  - `Internal\Framework\Module\Setup\Handler\ModuleConfigurationHandlerInterface`
+  - `Internal\Framework\Module\Setup\Service\ModuleConfigurationHandlingServiceInterface`
   - `Internal\Framework\Smarty\Module\TemplateExtension\TemplateBlockLoaderBridgeInterface`
   - `Internal\Framework\Templating\Loader\TemplateLoaderInterface`
-  - `Internal\Framework\Module\Setup\Handler\ModuleConfigurationHandlerInterface`
 - Classes:
   - `Application\Model\SmartyRenderer`
   - `Core\Module\ModuleSmartyPluginDirectoryRepository`
@@ -70,8 +71,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `oxfunctions::ox_get_template()`
   - `oxfunctions::ox_get_timestamp()`
   - `oxfunctions::ox_get_trusted()`
-  - `Internal\Framework\Module\Setup\Service\ModuleConfigurationHandlingService::handleOnDeactivation()`
-  - `Internal\Framework\Module\Setup\Service\ModuleConfigurationHandlingService::addHandler()`
 - Properties:
   - `Application\Controller\Admin\AdminDetailsController::$_oEditor`
   - `Core\UtilsView::$_aTemplateDir`